+++ test/zxspectrum/TestZ80Interface.h
@@ -45,8 +45,8 @@ namespace epoch::zxspectrum
             m_ram[address] = value;
         }
 
-        uint8_t ioRead(uint8_t port) override { return 0; }
-        void ioWrite(uint8_t port, uint8_t value) override {}
+        uint8_t ioRead(uint16_t port) override { return 0; }
+        void ioWrite(uint16_t port, uint8_t value) override {}
 
         [[nodiscard]] std::span<uint8_t> ram() { return m_ram; }
         [[nodiscard]] uint8_t ram(const uint16_t address) const { return m_ram[address]; }
